# api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Todo API
  version: 1.0.0
  description: API for the Reactive Todo Application
servers:
  - url: http://localhost:3000
paths:
  /todos:
    get:
      summary: Get all todos
      operationId: getTodos
      responses:
        '200':
          description: List of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
    post:
      summary: Create a new todo
      operationId: createTodo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  minLength: 1
                targetDate:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Created todo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'

  /todos/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: ID of the todo
    patch:
      summary: Update a todo
      operationId: updateTodo
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                isCompleted:
                  type: boolean
                targetDate:
                  type: string
                  format: date-time
                  nullable: true
                completedBy:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Updated todo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a todo
      operationId: deleteTodo
      responses:
        '200':
          description: Deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Todo:
      type: object
      required:
        - id
        - title
        - isCompleted
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        title:
          type: string
        isCompleted:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        targetDate:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        completedBy:
          type: string
          nullable: true
    Error:
      type: object
      properties:
        message:
          type: string